[ADD] removing node when env var not found
