+++ Makefile
@@ -32,6 +32,8 @@ SRCS			:=	main.c \
 SRCS_PARSER		:=	parse_args.c \
 					parse_pipe.c \
 					syntax_rules.c \
+					cmd_utils.c \
+					red_utils.c \
 
 # Lexer srcs
 SRCS_LEXER		:=	delimit_token.c \
@@ -40,6 +42,10 @@ SRCS_LEXER		:=	delimit_token.c \
 					lex_word.c \
 					lex_quote.c \
 					state_type.c \
+					token_add.c \
+					token_find.c \
+					token_free.c \
+					token_remove.c \
 
 # Expander srcs
 SRCS_EXPANDER	:=	expand_utils.c \
@@ -99,11 +105,6 @@ SRCS_UTILS		:=	exits.c \
 					ft_strncmp.c \
 					ft_strndup.c \
 					ft_substr.c \
-					token_append.c \
-					token_last.c \
-					token_clear.c \
-					token_remove.c \
-					token_new.c \
 					malloc_check.c \
 					strjoin_free.c \
 					test_list.c \
