+++ srcs/main.c
@@ -15,6 +15,7 @@ void	enter_shell(char **argv, char **envp)
 		{
 			token_printHtT(data.lexer.tokens);
 			expander(data.envp, &data.lexer);
+			token_printHtT(data.lexer.tokens);
 			if (!parser(data.lexer.tokens, &data.cmd))
 			{
 				g_state = EXECUTING;
